update prediction
Jagruti, I put the data files on my github so that we can now read directly from there.  I changed the names of the variables (t,u,v,w) so that they are more meaningful.  I hope that isn't too annoying for you.  I wanted the professors to be able to follow our work.  I also changed the titles of the columns because a number of them didn't mean very much to me.  I took out fileEncoding="latin1" even though I think it is important because I got an error every time I ran the code with that. I would like to put it back in. I also changed your merge statement because that wouldn't run for me, and I made the join an outer join (and not an inner join) so that we didn't lose tons of rows.